requirements-dev.txt

black>=24.0
pylint==2.7.1
qiskit-sphinx-theme>=1.6
sphinx-autodoc-typehints
Sphinx>=1.8.3
stestr>=2.0.0,!=4.0.0